Parameter To_other_map.1-A

This module is basically "Applicative", but with an additional type that can be used to translate to the applicative type.

Typically these types will be something like

type 'a t = 'a Value.t
type 'a s = 'a Form.t

val translate: 'a Form.t Value.t -> 'a Form.t Value.t

translate probably doesn't need to do any work. It just exists to expose a type equality between 'a Data.t and 'a s t.

type 'a t

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

val all : 'a t list -> 'a list t

type 'a s

val translate : 'a Tree.t -> 'a s t